sdk/tests/fast/canvas-test-expected.txt

This test ensures WebGL implementations interact correctly with the canvas tag.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Canvas.getContext
PASS context exists

Checking canvas and WebGL interaction
PASS canvas.width is 300
PASS canvas.height is 150
PASS pixels are 0,0,0,0
FAIL getViewport() should be 0,0,300,150. Was 0,0,0,0.

change display size of canvas and see that viewport does not change
 
